added migrations

+++ Presentation.Web/Controllers/API/AdviceUserRelationController.cs
@@ -5,18 +5,25 @@ using System;
 using System.Linq;
 using System.Net.Http;
 using System.Web.Mvc;
+using Core.ApplicationServices.Authorization;
 using Presentation.Web.Infrastructure.Attributes;
 
 namespace Presentation.Web.Controllers.API
 {
     [InternalApi]
+    [MigratedToNewAuthorizationContext]
     public class AdviceUserRelationController : GenericApiController<AdviceUserRelation, AdviceUserRelationDTO>
     {
-        IGenericRepository<AdviceUserRelation> _repository;
-        public AdviceUserRelationController(IGenericRepository<AdviceUserRelation> repository) : base(repository)
+        private readonly IGenericRepository<AdviceUserRelation> _repository;
+
+        public AdviceUserRelationController(
+            IGenericRepository<AdviceUserRelation> repository, 
+            IAuthorizationContext authorizationContext)
+            : base(repository, authorizationContext)
         {
             _repository = repository;
         }
+
         /// <summary>
         /// Sletter adviser med det specificerede id fra en genereisk advis
         /// </summary>
@@ -27,10 +34,18 @@ namespace Presentation.Web.Controllers.API
         {
             try
             {
-                foreach (var d in _repository.AsQueryable().Where(d => d.AdviceId == adviceId)) {
-                    _repository.Delete(d);
+                foreach (var d in _repository.AsQueryable().Where(d => d.AdviceId == adviceId))
+                {
+                    if (AllowDelete(d))
+                    {
+                        _repository.Delete(d);
+                        _repository.Save();
+                    }
+                    else
+                    {
+                        return Forbidden();
+                    }
                 }
-                _repository.Save();
                 return Ok();
             }
             catch (Exception e)
